+++ magpie/api/management/user/user_views.py
@@ -105,7 +105,7 @@ def delete_user_view(request):
                      http_error=HTTPForbidden, msg_on_fail=s.User_DELETE_ForbiddenResponseSchema.description)
 
     # Process any webhook requests
-    webhook_params = {"user_name": user.user_name, "user_id": user.id, "user_email": user.email}
+    webhook_params = {"user.name": user.user_name, "user.id": user.id, "user.email": user.email}
     process_webhook_requests(WebhookAction.DELETE_USER, webhook_params)
 
     return ax.valid_http(http_success=HTTPOk, detail=s.User_DELETE_OkResponseSchema.description)